+++ src/device/bot.ts
@@ -49,12 +49,11 @@ export class Bot {
   // Config
   botMode!: string;
   allowPush?: boolean;
+  doublePress!: number;
+  pushRatePress!: number;
   scanDuration!: number;
   deviceLogging!: string;
   deviceRefreshRate!: number;
-
-  // Others
-  doublePress!: number;
 
   // Updates
   botUpdateInProgress!: boolean;
@@ -346,11 +345,15 @@ export class Bot {
       )
       .subscribe(async () => {
         try {
-          interval(15000)
-            .pipe(take(this.doublePress!))
-            .subscribe(async () => {
-              await this.pushChanges();
-            });
+          if (this.doublePress > 1) {
+            interval(this.pushRatePress * 1000)
+              .pipe(take(this.doublePress!))
+              .subscribe(async () => {
+                await this.pushChanges();
+              });
+          } else {
+            await this.pushChanges();
+          }
         } catch (e: any) {
           this.apiError(e);
           this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} failed pushChanges with ${this.device.connectionType} Connection,`
@@ -364,7 +367,9 @@ export class Bot {
    * Parse the device status from the SwitchBot api
    */
   async parseStatus(): Promise<void> {
-    if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} parseStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else if (this.BLE) {
       await this.BLEparseStatus();
     } else if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIparseStatus();
@@ -419,7 +424,9 @@ export class Bot {
    * Asks the SwitchBot API for the latest device information
    */
   async refreshStatus(): Promise<void> {
-    if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} refreshStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else if (this.BLE) {
       await this.BLERefreshStatus();
     } else if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIRefreshStatus();
@@ -554,7 +561,9 @@ export class Bot {
    * Bot   -    "command"     "press"     "default"	  =        trigger press
    */
   async pushChanges(): Promise<void> {
-    if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} pushChanges enableCloudService: ${this.device.enableCloudService}`);
+    } else if (this.BLE) {
       await this.BLEpushChanges();
     } else if (this.OpenAPI) {
       await this.openAPIpushChanges();
@@ -562,7 +571,9 @@ export class Bot {
       this.debugWarnLog(`${this.device.deviceType}: ${this.accessory.displayName} Connection Type:`
       + ` ${this.device.connectionType}, pushChanges will not happen.`);
     }
-    interval(5000)
+    // Refresh the status from the API
+    interval(15000)
+      .pipe(skipWhile(() => this.botUpdateInProgress))
       .pipe(take(1))
       .subscribe(async () => {
         await this.refreshStatus();
@@ -1289,6 +1300,14 @@ export class Bot {
     } else if (this.platform.config.options!.refreshRate) {
       this.deviceRefreshRate = this.accessory.context.refreshRate = this.platform.config.options!.refreshRate;
       this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName} Using Platform Config refreshRate: ${this.deviceRefreshRate}`);
+    }
+    // pushRatePress
+    if (device?.bot?.pushRatePress) {
+      this.pushRatePress = device?.bot?.pushRatePress;
+      this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName} Using Device Config Bot pushRatePress: ${this.pushRatePress}`);
+    } else {
+      this.pushRatePress = 15;
+      this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName} Using Default Bot pushRatePress: ${this.pushRatePress}`);
     }
   }