+++ src/device/ceilinglight.ts
@@ -206,13 +206,6 @@ export class CeilingLight {
               + ` Error Message: ${superStringify(e.message)}`);
         }
         this.ceilingLightUpdateInProgress = false;
-        // Refresh the status from the API
-        interval(15000)
-          .pipe(skipWhile(() => this.ceilingLightUpdateInProgress))
-          .pipe(take(1))
-          .subscribe(async () => {
-            await this.refreshStatus();
-          });
       });
   }
 
@@ -232,7 +225,9 @@ export class CeilingLight {
    * Parse the device status from the SwitchBot api
    */
   async parseStatus(): Promise<void> {
-    /*if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} parseStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else /*if (this.BLE) {
       await this.BLEparseStatus();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIparseStatus();
@@ -307,7 +302,9 @@ export class CeilingLight {
    * Asks the SwitchBot API for the latest device information
    */
   async refreshStatus(): Promise<void> {
-    /*if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} refreshStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else/*if (this.BLE) {
       await this.BLERefreshStatus();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIRefreshStatus();
@@ -456,7 +453,9 @@ export class CeilingLight {
    *
    */
   async pushChanges(): Promise<void> {
-    /*if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} pushChanges enableCloudService: ${this.device.enableCloudService}`);
+    } else /*if (this.BLE) {
       await this.BLEpushChanges();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIpushChanges();
@@ -464,7 +463,9 @@ export class CeilingLight {
       this.debugWarnLog(`${this.device.deviceType}: ${this.accessory.displayName} Connection Type:`
       + ` ${this.device.connectionType}, pushChanges will not happen.`);
     }
-    interval(5000)
+    // Refresh the status from the API
+    interval(15000)
+      .pipe(skipWhile(() => this.ceilingLightUpdateInProgress))
       .pipe(take(1))
       .subscribe(async () => {
         await this.refreshStatus();