+++ src/device/colorbulb.ts
@@ -206,13 +206,6 @@ export class ColorBulb {
               + ` Error Message: ${superStringify(e.message)}`);
         }
         this.colorBulbUpdateInProgress = false;
-        // Refresh the status from the API
-        interval(15000)
-          .pipe(skipWhile(() => this.colorBulbUpdateInProgress))
-          .pipe(take(1))
-          .subscribe(async () => {
-            await this.refreshStatus();
-          });
       });
   }
 
@@ -220,7 +213,9 @@ export class ColorBulb {
    * Parse the device status from the SwitchBot api
    */
   async parseStatus(): Promise<void> {
-    /*if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} parseStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else/*if (this.BLE) {
       await this.BLEparseStatus();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIparseStatus();
@@ -295,7 +290,9 @@ export class ColorBulb {
    * Asks the SwitchBot API for the latest device information
    */
   async refreshStatus(): Promise<void> {
-    /*if (this.BLE) {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} refreshStatus enableCloudService: ${this.device.enableCloudService}`);
+    } else/*if (this.BLE) {
       await this.BLERefreshStatus();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
       await this.openAPIRefreshStatus();
@@ -444,6 +441,9 @@ export class ColorBulb {
    *
    */
   async pushChanges(): Promise<void> {
+    if (!this.device.enableCloudService && this.OpenAPI) {
+      this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName} pushChanges enableCloudService: ${this.device.enableCloudService}`);
+    } else
     /*if (this.BLE) {
       await this.BLEpushChanges();
     } else*/ if (this.OpenAPI && this.platform.config.credentials?.token) {
@@ -452,7 +452,9 @@ export class ColorBulb {
       this.debugWarnLog(`${this.device.deviceType}: ${this.accessory.displayName} Connection Type:`
       + ` ${this.device.connectionType}, pushChanges will not happen.`);
     }
-    interval(5000)
+    // Refresh the status from the API
+    interval(15000)
+      .pipe(skipWhile(() => this.colorBulbUpdateInProgress))
       .pipe(take(1))
       .subscribe(async () => {
         await this.refreshStatus();