+++ src/irdevice/airpurifier.ts
@@ -127,21 +127,25 @@ export class AirPurifier {
    */
   async pushAirPurifierOnChanges(): Promise<void> {
     if (this.Active !== 1) {
+      const commandType: string = await this.commandType();
+      const command: string = await this.commandOn();
       const body = superStringify({
-        'command': 'turnOn',
+        'command': command,
         'parameter': 'default',
-        'commandType': 'command',
+        'commandType': commandType,
       });
       await this.pushChanges(body);
     }
   }
 
   async pushAirPurifierOffChanges(): Promise<void> {
+    const commandType: string = await this.commandType();
+    const command: string = await this.commandOff();
     if (this.Active !== 0) {
       const body = superStringify({
-        'command': 'turnOff',
+        'command': command,
         'parameter': 'default',
-        'commandType': 'command',
+        'commandType': commandType,
       });
       await this.pushChanges(body);
     }
@@ -261,6 +265,36 @@ export class AirPurifier {
         + ` updateCharacteristic CurrentHeaterCoolerState: ${this.CurrentHeaterCoolerState}`,
       );
     }
+  }
+
+  async commandType() {
+    let commandType: string;
+    if (this.device.customize) {
+      commandType = 'customize';
+    } else {
+      commandType = 'command';
+    }
+    return commandType;
+  }
+
+  async commandOn() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOn';
+    }
+    return command;
+  }
+
+  async commandOff() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOff';
+    }
+    return command;
   }
 
   async statusCode({ res }: { res: IncomingMessage }): Promise<void> {