+++ src/irdevice/tv.ts
@@ -221,20 +221,24 @@ export class TV {
    */
   async pushTvOnChanges(): Promise<void> {
     if (this.Active !== 1) {
+      const commandType: string = await this.commandType();
+      const command: string = await this.commandOn();
       const body = superStringify({
-        'command': 'turnOn',
+        'command': command,
         'parameter': 'default',
-        'commandType': 'command',
+        'commandType': commandType,
       });
       await this.pushTVChanges(body);
     }
   }
 
   async pushTvOffChanges(): Promise<void> {
+    const commandType: string = await this.commandType();
+    const command: string = await this.commandOff();
     const body = superStringify({
-      'command': 'turnOff',
+      'command': command,
       'parameter': 'default',
-      'commandType': 'command',
+      'commandType': commandType,
     });
     await this.pushTVChanges(body);
   }
@@ -390,6 +394,36 @@ export class TV {
       this.tvService?.updateCharacteristic(this.platform.Characteristic.ActiveIdentifier, this.ActiveIdentifier);
       this.debugLog(`${this.device.remoteType}: ${this.accessory.displayName}` + ` updateCharacteristic ActiveIdentifier: ${this.ActiveIdentifier}`);
     }
+  }
+
+  async commandType() {
+    let commandType: string;
+    if (this.device.customize) {
+      commandType = 'customize';
+    } else {
+      commandType = 'command';
+    }
+    return commandType;
+  }
+
+  async commandOn() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOn';
+    }
+    return command;
+  }
+
+  async commandOff() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOff';
+    }
+    return command;
   }
 
   async statusCode({ res }: { res: IncomingMessage }): Promise<void> {