+++ src/irdevice/waterheater.ts
@@ -78,10 +78,12 @@ export class WaterHeater {
    */
   async pushWaterHeaterOnChanges(): Promise<void> {
     if (this.Active !== 1) {
+      const commandType: string = await this.commandType();
+      const command: string = await this.commandOn();
       const body = superStringify({
-        'command': 'turnOn',
+        'command': command,
         'parameter': 'default',
-        'commandType': 'command',
+        'commandType': commandType,
       });
       await this.pushChanges(body);
     }
@@ -89,10 +91,12 @@ export class WaterHeater {
 
   async pushWaterHeaterOffChanges(): Promise<void> {
     if (this.Active !== 0) {
+      const commandType: string = await this.commandType();
+      const command: string = await this.commandOff();
       const body = superStringify({
-        'command': 'turnOff',
+        'command': command,
         'parameter': 'default',
-        'commandType': 'command',
+        'commandType': commandType,
       });
       await this.pushChanges(body);
     }
@@ -158,6 +162,36 @@ export class WaterHeater {
       this.valveService?.updateCharacteristic(this.platform.Characteristic.Active, this.Active);
       this.debugLog(`${this.device.remoteType}: ${this.accessory.displayName} updateCharacteristic Active: ${this.Active}`);
     }
+  }
+
+  async commandType() {
+    let commandType: string;
+    if (this.device.customize) {
+      commandType = 'customize';
+    } else {
+      commandType = 'command';
+    }
+    return commandType;
+  }
+
+  async commandOn() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOn';
+    }
+    return command;
+  }
+
+  async commandOff() {
+    let command: string;
+    if (this.device.customize && this.device.customOn) {
+      command = this.device.customOn;
+    } else {
+      command = 'turnOff';
+    }
+    return command;
   }
 
   async statusCode({ res }: { res: IncomingMessage }): Promise<void> {